refactor: renames (de)serialize to (de)serialise
