+++ lib/fetch.js
@@ -3,7 +3,7 @@ module.exports = fetch
 var request = require('../helpers/request')
 var saveSession = require('../helpers/save-session')
 var get = require('./get')
-var deserialize = require('../helpers/deserialize')
+var deserialise = require('../helpers/deserialise')
 
 function fetch (state, basePath, path) {
   return request({
@@ -17,7 +17,7 @@ function fetch (state, basePath, path) {
 
   .then(function (response) {
     var data = JSON.parse(response.body)
-    data = deserialize(data)
+    data = deserialise(data)
 
     // split the basePath into an array and grab the last item
     var type = basePath.split('.')
