+++ lib/signin.js
@@ -3,8 +3,8 @@ module.exports = signIn
 var request = require('../helpers/request')
 var Promise = require('../helpers/promise')
 var saveSession = require('../helpers/save-session')
-var deserialize = require('../helpers/deserialize')
-var serialize = require('../helpers/serialize')
+var deserialise = require('../helpers/deserialise')
+var serialise = require('../helpers/serialise')
 
 function signIn (state, options) {
   if (!options || !options.username || !options.password) {
@@ -18,12 +18,12 @@ function signIn (state, options) {
       'Content-Type': 'application/vnd.api+json',
       'Accept': 'application/vnd.api+json'
     },
-    body: JSON.stringify(serialize('session', options))
+    body: JSON.stringify(serialise('session', options))
   })
 
   .then(function (response) {
     var data = JSON.parse(response.body)
-    data = deserialize(data)
+    data = deserialise(data)
 
     saveSession(state, { id: data.id })
     saveSession(state, { username: options.username }, 'account')
