+++ lib/signup.js
@@ -3,8 +3,8 @@ module.exports = signUp
 var request = require('../helpers/request')
 var Promise = require('../helpers/promise')
 var saveSession = require('../helpers/save-session')
-var deserialize = require('../helpers/deserialize')
-var serialize = require('../helpers/serialize')
+var deserialise = require('../helpers/deserialise')
+var serialise = require('../helpers/serialise')
 
 function signUp (state, options) {
   if (!options || !options.username || !options.password) {
@@ -24,12 +24,12 @@ function signUp (state, options) {
       'Content-Type': 'application/vnd.api+json',
       'Accept': 'application/vnd.api+json'
     },
-    body: JSON.stringify(serialize('account', options))
+    body: JSON.stringify(serialise('account', options))
   })
 
   .then(function (response) {
     var data = JSON.parse(response.body)
-    data = deserialize(data, {
+    data = deserialise(data, {
       relationships: ['profile']
     })
