+++ lib/update-profile.js
@@ -3,7 +3,7 @@ module.exports = updateProfile
 var request = require('../helpers/request')
 var Promise = require('../helpers/promise')
 var saveSession = require('../helpers/save-session')
-var serialize = require('../helpers/serialize')
+var serialise = require('../helpers/serialise')
 
 function updateProfile (state, options) {
   if (!options) {
@@ -18,7 +18,7 @@ function updateProfile (state, options) {
       'Accept': 'application/vnd.api+json',
       'Authorization': 'Bearer ' + state.session.id
     },
-    body: JSON.stringify(serialize('profile', options, state.session.account.profile.id))
+    body: JSON.stringify(serialise('profile', options, state.session.account.profile.id))
   })
 
   .then(function () {
