+++ tests/index.js
@@ -8,5 +8,5 @@ require('./specs/fetch-profile')
 require('./specs/get')
 require('./specs/get-profile')
 require('./specs/update-profile')
-require('./specs/deserialize')
-require('./specs/serialize')
+require('./specs/deserialise')
+require('./specs/serialise')
